requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
newspaper3k
spacy
sentence-transformers
scikit-learn
pillow
torch
transformers
python-multipart
aiofiles
